[Refactor] : '도형(객체) 편집' 기능개발 #14 - 프로젝트 폴더 구조 개선
1. 훅 폴더 구조 개선
   - 캔버스 관련 훅을 `/hooks/canvas` 디렉토리로 이동
   - 각 훅 파일의 import 경로 수정 및 확장자 명시

2. 상태 관리 분리
   - `useCanvasRemoteStore`를 `/store` 디렉토리로 이동
   - 상태 관리 관련 타입 및 훅을 중앙화된 위치로 구성
