+++ src/hooks/canvas/useCanvasElementManager.ts
