+++ src/hooks/canvas/usePaintingCanvas.ts
@@ -1,10 +1,10 @@
 import * as React from 'react';
-import { ElementRegistry } from '@/hooks/useCanvasElementManager.ts';
-import { ViewManagerState } from '@/hooks/useCanvasViewManager.ts';
+import { ElementRegistry } from '@/hooks/canvas/useCanvasElementManager.ts';
+import { ViewManagerState } from '@/hooks/canvas/useCanvasViewManager.ts';
 import { useEffect } from 'react';
-import { SelectManagerState } from '@/hooks/useCanvasSelectManager.ts';
+import { SelectManagerState } from '@/hooks/canvas/useCanvasSelectManager.ts';
 import { colorToken } from '@/style/color';
-import { CreateElementManagerState } from '@/hooks/useCanvasCreateElementManger.ts';
+import { CreateElementManagerState } from '@/hooks/canvas/useCanvasCreateElementManger.ts';
 
 /**
  * ### usePaintingCanvas()
