+++ src/store/index.ts
@@ -1,0 +1,2 @@
+export { useCanvasRemoteStore } from './useCanvasRemoteStore';
+export type { ShapeType, RemoteMode } from './useCanvasRemoteStore';
